More logging, moved some UI around, preparing for manifest V3, automatic actor mod spec

+++ C2GUILauncher/src/FileHelpers.cs
@@ -1,10 +1,15 @@
-﻿using System;
+﻿using log4net;
+using log4net.Repository.Hierarchy;
+using System;
 using System.Collections.Generic;
+using System.Diagnostics.Eventing.Reader;
 using System.IO;
 using System.Linq;
 
 namespace C2GUILauncher {
     static class FileHelpers {
+        private static readonly ILog logger = LogManager.GetLogger(nameof(FileHelpers));
+
         public const string USER_LOCK_SUFFIX = ".USER_LOCK";
 
         public static bool IsFileLocked(string filePath) {
@@ -13,9 +18,12 @@ namespace C2GUILauncher {
         }
 
         public static bool DeleteFiles(IEnumerable<string> filePaths) {
+            logger.Info("Deleting files: " + string.Join("\n    ", filePaths));
             // Delete all files in a transaction.
             // If any of them are locked, don't delete any of them.
             if (filePaths.Any(IsFileLocked)) {
+                var lockedFiles = filePaths.Where(IsFileLocked);
+                logger.Warn($"Not deleting files because some of them are locked:\n" + string.Join("\n    ", lockedFiles));
                 return false;
             }
 
@@ -28,11 +36,32 @@ namespace C2GUILauncher {
 
         public static bool DeleteFile(string filePath) {
             if (IsFileLocked(filePath)) {
+                logger.Warn($"Not deleting file because it is locked: {filePath}");
                 return false;
             }
 
-            File.Delete(filePath);
+            if (File.Exists(filePath)) {
+                logger.Info("Deleting file: " + filePath);
+                File.Delete(filePath);
+            }
+
             return true;
+        }
+
+        public static bool DeleteDirectory(string filePath) {
+            logger.Info("Deleting directory: " + filePath);
+            if(!Directory.Exists(filePath)) {
+                logger.Warn($"Not deleting directory because it doesn't exist: {filePath}");
+                return false;
+            }
+
+            var files = Directory.GetFiles(filePath, "*", SearchOption.AllDirectories);
+            var result = DeleteFiles(files);
+
+            if(result)
+                Directory.Delete(filePath, true);
+
+            return result;
         }
 
         // Currently unused, but maybe someday.
